Se ajustaron detalles en comentarios y limpieza del codigo

+++ src/components/Contactos.jsx
@@ -1,23 +1,33 @@
+//--> Start Imports
+
 //Import Components
 import CardContacto from "./CardContacto";
 
 //Import Libraries
 import { useState, useEffect } from "react";
 import { collection, getDocs, doc, deleteDoc } from "firebase/firestore";
-
-//Import firestore
+import swal from "sweetalert";
 import db from "../firebase/firebaseConfig";
 
 //Import Hooks
 import { useContactContext, useContactToggleContext } from "../UseProvider";
 
+//--> End Imports
+
 const Contactos = ({ setEditionContact, setModeEdition }) => {
-	//States
+	//-->Start States
+
+	//contactos guarda los contactos en un arreglo de objetos y los muestra en pantalla
 	const [contactos, setContactos] = useState([]);
+	//stateContact cambia true o false cada que se modifica, se elimina y se guarda un contacto nuevo, sirve para que el useEffect este actualizando la lista de contactos.
 	const stateContact = useContactContext();
+
+	//changeStateContact es una funcion que sirve para modificar el valor del stateContact
 	const changeStateContact = useContactToggleContext();
 
-	//Consultar la base de datos de contactos en firestore
+	//-->End States
+
+	//Funcion que trae todos los contactos que actualmente estan en la base de datos firestore
 	const agendaCollection = collection(db, "agenda");
 	const queryDataAgenda = async () => {
 		const query = await getDocs(agendaCollection);
@@ -25,7 +35,7 @@ const Contactos = ({ setEditionContact, setModeEdition }) => {
 			...doc.data(),
 			id: doc.id,
 		}));
-		//Order contacts by name
+		//Ordenar contactos por nombre
 		function sortArray(x, y) {
 			if (x.nombre < y.nombre) {
 				return -1;
@@ -39,20 +49,41 @@ const Contactos = ({ setEditionContact, setModeEdition }) => {
 		setContactos(dataContactsOrder);
 	};
 
-	//Llamar a la funcion que consulta la base de datos
+	//UseEffect que se renderiza cada que el stateContact cambia y trae la lista actualizada de los contactos.
 	useEffect(() => {
 		queryDataAgenda();
 	}, [stateContact]);
 
-	//Eliminar contacto
+	//Funcion que elimina el contacto seleccionado
 	const deleteContact = async (id) => {
-		const contactDoc = doc(db, "agenda", id);
-		await deleteDoc(contactDoc);
-		queryDataAgenda();
+		let buttonDelete = false;
+		await swal({
+			title: "Eliminar Contacto",
+			text: "¿Estas seguro que deseas eliminar este contacto?",
+			icon: "warning",
+			buttons: true,
+			dangerMode: true,
+		}).then((willDelete) => {
+			if (willDelete) {
+				swal("Contacto eliminado", {
+					icon: "success",
+				});
+				buttonDelete = true;
+			} else {
+				buttonDelete = false;
+			}
+		});
+		console.log(buttonDelete);
+
+		if (buttonDelete === true) {
+			const contactDoc = doc(db, "agenda", id);
+			await deleteDoc(contactDoc);
+			queryDataAgenda();
+		}
 		changeStateContact();
 	};
 
-	//Editar contacto
+	//Funcion que activa el modo de edicion y filtra de la lista de contactos, el contacto que se va a editar
 	const editContact = (id) => {
 		const filterContact = contactos.filter((item) => item.id === id);
 		setEditionContact(filterContact);
